requirements.txt
# Core dependencies
langchain>=0.0.267
pypdf>=3.15.1
chromadb>=0.4.13
sentence-transformers>=2.2.2
openai>=1.2.0
tiktoken>=0.5.1
pydantic>=2.3.0
fastapi>=0.103.1
uvicorn>=0.23.2

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.12.0
webdriver-manager>=4.0.0

# Sanskrit processing
indic-nlp-library>=0.91
pyiwn>=0.0.5

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Scheduling and async
schedule>=1.2.0
asyncio>=3.4.3

# Cloud integration
boto3>=1.28.38  # For AWS
google-cloud-storage>=2.10.0  # For GCP
azure-storage-blob>=12.17.0  # For Azure

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
loguru>=0.7.0

# API and server
flask>=2.3.0
flask-cors>=4.0.0